+++ Kernel/Modules/AdminSystemAddress.pm
@@ -276,8 +276,21 @@ sub _Edit {
     $LayoutObject->Block( Name => 'ActionList' );
     $LayoutObject->Block( Name => 'ActionOverview' );
 
-    # get valid list
-    my %ValidList        = $Kernel::OM->Get('Kernel::System::Valid')->ValidList();
+    # Get valid list.
+    my $ValidObject = $Kernel::OM->Get('Kernel::System::Valid');
+    my %ValidList   = $ValidObject->ValidList();
+
+    # If there is queue using this system address, disable invalid selection on edit screen.
+    if ( $Param{Action} eq 'Change' ) {
+        $Param{SystemAddressIsUsed} = $Kernel::OM->Get('Kernel::System::SystemAddress')->SystemAddressIsUsed(
+            SystemAddressID => $Param{ID},
+        );
+        if ( $Param{SystemAddressIsUsed} ) {
+            my @ValidIDsList = $ValidObject->ValidIDsGet();
+            %ValidList = map { $_ => $ValidList{$_} } @ValidIDsList;
+        }
+    }
+
     my %ValidListReverse = reverse %ValidList;
 
     $Param{ValidOption} = $LayoutObject->BuildSelection(
@@ -302,7 +315,7 @@ sub _Edit {
         },
     );
 
-    # add the correct server error msg for the system email address
+    # Add the correct server error msg for the system email address.
     if ( $Param{Name} && $Param{Errors}->{ErrorType} ) {
         $LayoutObject->Block(
             Name => 'Email' . $Param{Errors}->{ErrorType} . 'ServerErrorMsg',